Added navigate action and refactorized
