+++ app/src/main/java/com/example/recipesapp/navigation/screens/HomeScreen.kt
@@ -33,7 +33,6 @@ import com.example.recipesapp.components.recipes.RecommendedRecipeList
 import com.example.recipesapp.ui.theme.LightBrown
 import com.example.recipesapp.components.recipes.SearchBar
 
-import androidx.compose.foundation.layout.*
 import androidx.compose.material.icons.Icons
 import androidx.compose.material.icons.filled.Favorite
 import androidx.compose.material.icons.filled.Home
@@ -41,7 +40,6 @@ import androidx.compose.material.icons.filled.Settings
 
 import androidx.compose.material3.Icon
 import androidx.compose.material3.BottomAppBar
-import androidx.compose.material3.BottomAppBarDefaults
 import androidx.compose.material3.MaterialTheme
 import androidx.compose.material3.NavigationBarItem
 import androidx.compose.runtime.Stable
@@ -54,14 +52,13 @@ import androidx.compose.ui.res.stringResource
 import androidx.compose.ui.tooling.preview.Preview
 import androidx.compose.ui.unit.DpOffset
 import com.example.recipesapp.components.common.BellColorButton
-import com.example.recipesapp.ui.theme.DarkBrown
-import com.example.recipesapp.ui.theme.LightPurple
+import com.example.recipesapp.components.common.BottomNavigation
+import com.example.recipesapp.ui.theme.LightGray
 import com.example.recipesapp.ui.theme.MediumBrown
-import com.example.recipesapp.ui.theme.RoyalPurple
+import com.example.recipesapp.ui.theme.RoyalGray
 import com.exyte.animatednavbar.AnimatedNavigationBar
 import com.exyte.animatednavbar.animation.balltrajectory.Teleport
 import com.exyte.animatednavbar.animation.indendshape.Height
-import com.exyte.animatednavbar.animation.indendshape.shapeCornerRadius
 import com.exyte.animatednavbar.items.wigglebutton.WiggleButton
 
 
@@ -74,15 +71,13 @@ fun HomeScreen(
 ) {
     val viewModel: RecipeViewModel = hiltViewModel()
     val uiState = viewModel.state.collectAsState()
-    var selectedItem by remember { mutableStateOf(0) }
     Scaffold(
         topBar = {
             SearchBar(placeholder = "Search recipes...",
                 action = { query -> viewModel.getSearchRecipe2(query) })
         },
         bottomBar = {
-            //SimpleBottomNavigationBar(navController)
-            BottomNav()
+            BottomNavigation(navController)
         }
     ) { innerPadding ->
         Column(
@@ -151,109 +146,9 @@ fun HomeScreen(
     }
 }
 
-
-@Composable
-fun SimpleBottomNavigationBar(navController: NavHostController) {
-    val currentRoute = navController.currentBackStackEntry?.destination?.route
-    BottomAppBar(
-        modifier = Modifier
-            .padding(horizontal = 16.dp, 16.dp)
-            .background(Color.Transparent)
-    ) {
-        ScreenEnum.values().forEach { screen ->
-            NavigationBarItem(
-                icon = { Icon(screen.icon, contentDescription = screen.title) },
-                label = { Text(screen.title) },
-                selected = currentRoute == screen.route,
-                onClick = { navController.navigate(screen.route) }
-            )
-        }
-    }
-}
-
 enum class ScreenEnum(val route: String, val icon: ImageVector, val title: String) {
     Home("home", Icons.Filled.Home, Screen.Home.title),
     Favorites("favorites", Icons.Filled.Favorite, "Favorites"),
     Settings("settings", Icons.Filled.Settings, "Settings")
 }
 
-@Stable
-data class WiggleButtonItem(
-    @DrawableRes val backgroundIcon: Int,
-    @DrawableRes val icon: Int,
-    var isSelected: Boolean,
-    @StringRes val description: Int,
-    val animationType: BellColorButton = BellColorButton(
-        tween(500),
-        background = ButtonBackground(R.drawable.plus)
-    ),
-)
-
-data class ButtonBackground(
-    @DrawableRes val icon: Int,
-    val offset: DpOffset = DpOffset.Zero
-)
-
-
-val wiggleButtonItems = listOf(
-    WiggleButtonItem(
-        icon = R.drawable.outline_home,
-        backgroundIcon = R.drawable.home,
-        isSelected = false,
-        description = R.string.Home,
-    ),
-    WiggleButtonItem(
-        icon = R.drawable.outline_favorite,
-        backgroundIcon = R.drawable.favorite,
-        isSelected = false,
-        description = R.string.Heart
-    ),
-    WiggleButtonItem(
-        icon = R.drawable.outline_circle,
-        backgroundIcon = R.drawable.circle,
-        isSelected = false,
-        description = R.string.Circle
-    ),
-)
-
-const val Duration = 500
-const val DoubleDuration = 1000
-@Preview
-@Composable
-fun BottomNav() {
-    var selectedItem by remember { mutableStateOf(0) }
-
-        AnimatedNavigationBar(
-            modifier = Modifier
-                .height(85.dp)
-                .background(MaterialTheme.colorScheme.onSurface), // O hacer transparente el fondo del AnimatedNavigationBar
-            selectedIndex = selectedItem,
-            ballColor = Color.White,
-            ballAnimation = Teleport(tween(Duration, easing = LinearEasing)),
-            indentAnimation = Height(
-                indentWidth = 56.dp,
-                indentHeight = 15.dp,
-                animationSpec = tween(
-                    DoubleDuration,
-                    easing = { OvershootInterpolator().getInterpolation(it) })
-            )
-        ) {
-            wiggleButtonItems.forEachIndexed { index, it ->
-                WiggleButton(
-                    modifier = Modifier.fillMaxSize().background(MediumBrown), // Opción para hacer los botones también transparentes
-                    isSelected = selectedItem == index,
-                    onClick = { selectedItem = index },
-                    icon = it.icon,
-                    backgroundIcon = it.backgroundIcon,
-                    wiggleColor = LightPurple,
-                    outlineColor = RoyalPurple,
-                    contentDescription = stringResource(id = it.description),
-                    enterExitAnimationSpec = tween(
-                        durationMillis = Duration,
-                        easing = LinearEasing
-                    ),
-                    wiggleAnimationSpec = spring(dampingRatio = .45f, stiffness = 35f)
-                )
-            }
-    }
-}
