+++ app/src/main/java/com/example/recipesapp/ui/theme/Color.kt
@@ -12,11 +12,8 @@ val Pink40 = Color(0xFF7D5260)
 
 val LightBrown = Color(0xFFECE1D0)
 val MediumBrown = Color(0xFFF0E9E0)
-val DarkBrown = Color(0xFF5A3B0B)
-val YellowButton = Color(4294829183)
 val RecipeCard = Color(0xFFF8F1E8)
 val BorderRecipeCard = Color(0xFFE1E1E1)
 
-val Purple = Color(0xFF742DF6)
-val LightPurple = Color(0xFF707070)
-val RoyalPurple = Color(0xFF1F1F1F)
+val LightGray = Color(0xFF707070)
+val RoyalGray = Color(0xFF1F1F1F)
